moved all redux reducers and actions to appropriate folder
