--- src/app.css
@@ -1,7 +1,0 @@
-.switch-wrapper {
-  position: relative;
-}
-
-.switch-wrapper > div {
-  position: absolute;
-}
