--- src/components/app/appPageRedux.js
@@ -1,10 +1,0 @@
-const initialState = {
-	disabled: false
-};
-
-export default function(state = initialState, action){
-  switch (action.type) {
-    default:
-      return state;
-  }
-}
